split some translation units into more granular ones
